requirements.txt
# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Progress bars
tqdm==4.66.1

# Date parsing
python-dateutil==2.8.2

praw
google-generativeai
linkedin-api